backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt
django-filter
mysqlclient
python-dotenv
django-cors-headers
gunicorn